Getting started
Wanted to make my own note taking app for the practice and whatnot.

Started making the very basic functionality. Might need to be completely reworked. But for now it's possible to "add" a note. Which basically takes the value of the input element and adds it as the textnode to a new "p" element.

Obviously this is very basic stuff only.